requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
asyncpg
openai
pgvector
scikit-learn
hdbscan
numpy
python-dotenv
pytest
pydantic